==> analysis/1-fit-models.Rmd <==
---
output:
  word_document: default
  html_document: default
---

# Bayesian analysis of seagrass stressors 

```{r, echo=FALSE}
library(knitr)
opts_knit$set(root.dir = "..") 
```

```{r, echo = FALSE}
knitr::opts_chunk$set(
  collapse = TRUE,
  comment = "#>"
)
```

```{r, message=FALSE}
library(tidyverse)
library(rstanarm)
library(RColorBrewer)
library(assertthat)
library(ggsidekick) # devtools::install_github("seananderson/ggsidekick")
```

We'll create this folder in case it isn't already there. 

```{r}
dir.create("figs", showWarnings = FALSE)
dir.create("data/generated", showWarnings = FALSE)
```

Let's read in the data:

```{r}
d1 <- read_csv("data/mesocosm_data_1.csv", na = c("NA", "na"))
d2 <- read_csv("data/mesocosm_data_2.csv", na = c("NA", "na"))
d3 <- read_csv("data/mesocosm_data_3.csv", na = c("NA", "na"))
labels <- read_csv("data/mesocosm_figure_labels.csv") %>% 
  rename(response = label_data, label = label_pub, 
    figure_panel = order, notes = Notes) %>% 
  select(-notes)
names(d1) <- tolower(names(d1))
names(d2) <- tolower(names(d2))
names(d3) <- tolower(names(d3))

sd_ph <- sd(d3$ph)
mean_ph <- mean(d3$ph)
scale_ph <- function(x) (x - mean_ph) / (2 * sd_ph)

d1 <- d1 %>% select(-rep) %>% 
  gather(response, value, -ph, -nutrients)
d2 <- d2 %>% 
  gather(response, value, -ph, -nutrients)
d3 <- d3 %>% select(-barrel) %>% 
  gather(response, value, -ph, -nutrients)

d <- bind_rows(d1, d2) %>% bind_rows(d3)
d <- mutate(d, nutrients_text = nutrients,
  nutrients = as.numeric(as.factor(nutrients)) - 1)
assert_that(identical(unique(d$nutrients), c(0, 1)))

d <- left_join(d, labels)
assert_that(all(!is.na(d$label)))

# `ulva_grazing` is zero for all cases of no nutrients; don't model
d_ulva0 <- filter(d, response == "ulva_grazing" & value == 0 & nutrients == 0)
d <- filter(d, !(response == "ulva_grazing" & value == 0 & nutrients == 0))

d <- filter(d, response != "change_plant_mass_g_fw")
```

We need to subtract 20 from each shoot elongation response and manipulate some units so that coefficients are roughly on the same scale:

```{r}
d$value[d$response == "shoot_length"] <- d$value[d$response == "shoot_length"] - 20
d$value[d$response == "change_plant_mass_g_fw"] <- d$value[d$response == "change_plant_mass_g_fw"]/10
d$value[d$response == "shoot_length"] <- d$value[d$response == "shoot_length"]/10
d$value[d$response == "rhi_elong_cm"] <- d$value[d$response == "rhi_elong_cm"]/10
d <- filter(d, !is.na(value))
```

```{r}
ggplot(d, aes(ph, value, colour = nutrients_text)) + 
  geom_point(position = position_jitter(width = 0.03), alpha = 0.5) +
  facet_wrap(~response, scales = "free_y")
```

```{r, eval=FALSE, echo=FALSE}
all_whole_numbers <- function(x) {
  identical(x, round(x))
}

group_by(d, response) %>% 
  summarize(minimum_value = min(value), counts = all_whole_numbers(value)) %>% 
  filter(minimum_value == 0, !counts)

group_by(d, response) %>% 
  summarize(minimum_value = min(value), counts = all_whole_numbers(value)) %>% 
  filter(counts)

group_by(d, response) %>% 
  summarize(minimum_value = min(value), counts = all_whole_numbers(value)) %>% 
  filter(minimum_value < 0)
```

Should we condense the data in the eelgrass figure into tank means? 

```{r}
# d <- group_by(d, response, nutrients, ph, figure_panel, 
#   nutrients_text, label) %>% 
#   summarize(value = mean(value)) %>% 
#   ungroup()
```

Here I'm creating a data frame to merge in that describes the type of GLM to fit for each response variable:

```{r}
# link not used!
dt <- tribble(
  ~response,                  ~log_transform,  ~family,     ~link,      ~re,
  "algal_mass_g_dw",          TRUE,            "gaussian",  "identity", FALSE,
  "change_idotea_mass_g_fw",  FALSE,           "gaussian",  "identity", FALSE,
# "change_plant_mass_g_fw",   FALSE,           "gaussian",  "identity", FALSE,
  "change_seahare_mass_g_fw", FALSE,           "gaussian",  "identity", FALSE,
  "dead_shoots",              FALSE,           "count",     "log",      FALSE,
  "diatom_grazing",           TRUE,            "gaussian",  "identity", FALSE,
  "epiphyte_cm_shoot",        TRUE,            "gaussian",  "identity", FALSE,
  "ido_mortality",            FALSE,           "count",     "log",      FALSE,
  "rhi_elong_cm",             FALSE,           "gaussian",  "identity", TRUE,
  "rhi_mass_g_dw",            TRUE,            "gaussian",  "identity", TRUE,
  "sh_grazing",               TRUE,            "gaussian",  "identity", FALSE,
  "sh_mortality",             FALSE,           "count",     "log",      FALSE,
  "shoot_length",             FALSE,           "gaussian",  "identity", TRUE,
  "sht_g_dw",                 TRUE,            "gaussian",  "identity", FALSE,
  "ulva_grazing",             TRUE,            "gaussian",  "identity", FALSE)

assert_that(all(unique(d$response) %in% unique(dt$response)))
d_long <- left_join(d, dt)
meta <- unique(select(d_long, response, log_transform, family))
save(meta, file = "data/meta.rda")
```

Let's print a table that can be used in the paper to describe the models:

```{r}
.simpleCap <- function(x) {
  s <- strsplit(x, " ")[[1]]
  paste(toupper(substring(s, 1, 1)), substring(s, 2),
    sep = "", collapse = " ")
}

cap_vec <- function(x) {
  map_chr(x, .simpleCap)
}

model_table <- inner_join(dt, labels) %>% select(-response) %>% 
  mutate(figure_panel = toupper(figure_panel)) %>% 
  mutate(label = sub("\\([a-zA-Z0-9 /%]*\\)", "", label)) %>% 
  mutate(label = sub(" $", "", label)) %>% 
  mutate(label = sub("\\\\n", "", label)) %>% 
  mutate(family = sub("count", "Negative binomial", family)) %>% 
  mutate(family = cap_vec(family)) %>%
  mutate(link = cap_vec(link)) %>%
  arrange(figure_panel) %>%
  mutate(figure_panel = toupper(figure_panel)) %>% 
  select(figure_panel, label, log_transform, family, link, re) %>% 
  rename(`Random intercepts` = re)

names(model_table) <- cap_vec(names(model_table))
names(model_table) <- sub("_", " ", names(model_table))

knitr::kable(model_table)
```

Let's work with orthogonal polynomials:

```{r}
d_long$ph_scaled <- d$ph

d_long <- d_long %>% group_by(response) %>%
  mutate(
    ph_scaled = poly(ph, 2)[,1],
    ph_scaled2 = poly(ph, 2)[,2]) %>%
  ungroup()

sort(unique(filter(d_long, response == "diatom_grazing")$value))[2]
sort(unique(filter(d_long, response == "sh_grazing")$value))[2]
sort(unique(filter(d_long, response == "ulva_grazing")$value))[2]

d_long <- group_by(d_long, response) %>% 
  mutate(add_fudge = value == 0 & log_transform) %>% 
  mutate(value = ifelse(value == 0 & log_transform, sort(unique(value))[2]/4, value))
sum(d_long$add_fudge)
select(filter(d_long, add_fudge), ph, nutrients, response, value)
```

Let's plot the data that we will fit models to:

```{r, fig.width=8.5}
ggplot(d_long, aes(ph_scaled, value, colour = nutrients_text)) + 
  geom_point() +
  facet_wrap(~response, scales = "free_y") +
  theme_sleek()
```

Now we can iterate through the various response variables and fit a Bayesian GLM to each. For now, I am log transforming some of the response variables and fitting a linear regression, to others I am fitting a linear regression to the raw data (the change response variables), and I am fitting negative binomial GLMs to the count response variables.

```{r, message=FALSE, results='hide', cache=FALSE}
d_long <- mutate(d_long, tank = as.numeric(as.factor(paste(ph, nutrients))))
options(mc.cores = 1L)
if (!file.exists("data/generated/stan-fits.rds")) {
  fits <- plyr::dlply(d_long, "response", function(x) {
    print(unique(x$response))
    if (x$log_transform[1]) {
      f <- log(value) ~ ph_scaled + ph_scaled2 + nutrients
    } else
      f <- value ~ ph_scaled + ph_scaled2 + nutrients
    
    if (unique(x$response) == "ulva_grazing") # none when no nutrients
      f <- log(value) ~ ph_scaled + ph_scaled2
    
    if (x$family[1] == "gaussian") {
      fam <- gaussian(link = "identity")
      prior_aux <- student_t(df = 3, 0, 5, autoscale = FALSE)
    }
    #  if (x$family[1] == "gamma")
    #    fam <- gaussian(link = "identity")
    # fam <- Gamma(link = "log")
    if (x$family[1] == "count") {
      fam <- neg_binomial_2(link = "log")
      prior_aux <- student_t(df = 3, 0, 20, autoscale = FALSE)
    }
    
    if (unique(x$re) == FALSE) {
      stan_glm(f, data = x, iter = 2000, chains = 4, family = fam,
        prior = normal(0, 3, autoscale = FALSE),
        prior_intercept = normal(0, 10, autoscale = FALSE),
        prior_aux = prior_aux)
    } else {
      f <- as.formula(paste(Reduce(paste, deparse(f)), "+ (1 | tank)"))
      stan_lmer(f, data = x, iter = 2000, chains = 4,
        prior = normal(0, 3, autoscale = FALSE),
        prior_intercept = normal(0, 10, autoscale = FALSE),
        prior_aux = prior_aux, cores = 4)
    }
  })
  saveRDS(fits, file = "data/generated/stan-fits.rds")
} else {
  fits <- readRDS("data/generated/stan-fits.rds")
}
```

Let's check each of the models to make sure they converged:

```{r}
check_out <- map(fits, function(x) {
  assert_that(all(x$stan_summary[,"Rhat"] < 1.05))
  assert_that(all(x$stan_summary[,"n_eff"] > 200))
})
```

Now let's extract the samples from the posteriors and make them into a nice data frame to work with:

```{r}
fitsl <- map(fits, function(x) as.data.frame(as.matrix(x)))
fitsl$ulva_grazing$nutrients <- NA

s <- plyr::ldply(fitsl, function(x) {
  as_data_frame(select(x, ph_scaled, ph_scaled2, nutrients))
}) %>%
  rename(response = .id, linear = `ph_scaled`, quadratic = `ph_scaled2`)
```

```{r}
fits
```

Here I will calculate credible intervals to plot:

```{r}
cis <- gather(s, parameter, value, -response) %>% 
  group_by(response, parameter) %>% 
  summarise(l = quantile(value, 0.025, na.rm = TRUE),
    m = quantile(value, 0.5, na.rm = TRUE),
    u = quantile(value, 0.975, na.rm = TRUE)) %>% 
  ungroup() %>% 
  mutate(parameter = factor(parameter, levels = c("linear", "quadratic", "nutrients"))) %>% 
  left_join(labels)
```

Now let's plot the coefficients with the credible intervals overlaid:

```{r, fig.height=4, fig.width=6.5}
lims <- c(-5.2, 5.2)
cis$l[cis$l < lims[1]] <- lims[1]
cis$u[cis$u > lims[2]] <- lims[2]
cis <- mutate(cis, parameter = factor(parameter, levels = c("linear", "quadratic", "nutrients"),
  labels = c("Linear - pH", "Quadratic - pH", "Nutrients")))
cis$label <- sub("\\([a-zA-Z0-9 /%]*\\)", "", cis$label)
cis$label <- sub(" $", "", cis$label)
cis$label <- sub("\\\\n", "", cis$label)
cis <- mutate(cis, label_ordered = paste(label, toupper(figure_panel), sep = " - "))
cis <- mutate(cis, label_ordered = reorder(label_ordered, -as.numeric(as.factor(figure_panel))))
```

```{r}
cols <- brewer.pal(6, "Blues")
g <- gather(s, parameter, sample, -response) %>%
  inner_join(unique(select(d_long, response, log_transform, family, label, figure_panel))) %>%
  # mutate(multiplicative = ifelse(log_transform | family == "count", TRUE, FALSE)) %>%
  mutate(parameter = factor(parameter, levels = c("linear", "quadratic", "nutrients"),
    labels = c("Linear - pH", "Quadratic - pH", "Nutrients"))) %>%
  mutate(label = sub("\\([a-zA-Z0-9 /%]*\\)", "", label)) %>% 
  mutate(label = sub(" $", "", label)) %>% 
  mutate(label = sub("\\\\n", "", label)) %>% 
  mutate(label_ordered = paste(label, toupper(figure_panel), sep = " - ")) %>% 
  mutate(label_ordered = reorder(label_ordered, -as.numeric(as.factor(figure_panel)))) %>% 
  ggplot(aes(x = label_ordered, sample)) + 
  geom_hline(yintercept = 0, lty = 2, col = "grey60") +
  geom_violin(fill = paste0(cols[3]), col = NA, adjust = 1.2, 
    lwd = 0.1) +
  coord_flip() +
  facet_wrap(~parameter, scales = "free_x") +
  ylim(lims) +
  theme_sleek() +
  ylab("Coefficient") + xlab("Other        Grazing trials   Eelgrass  Algal  Grazer") +
  geom_point(data = cis, aes(x = label_ordered, y = m), col = "grey30", cex = 0.5) +
  geom_segment(data = cis, aes(x = label_ordered, y = l, xend = label_ordered, yend = u),
    lwd = 0.4, col = "grey30") +
  geom_vline(xintercept = c(3.5, 6.5, 10.5), col = "grey60", lty = 3)
g
ggsave("figs/fig-2.pdf", width = 7, height = 4)
```

What about quantifying the probability density above and below zero for each of these variables?

```{r}
gather(s, coefficient, value, -response) %>%
  group_by(response, coefficient) %>% 
  summarize(p_less_0 = round(sum(value < 0)/n(), 2), p_above_0 = round(1 - p_less_0, 2)) %>%
  left_join(labels) %>% 
  mutate(figure_number = substr(figure_panel, 0, 1)) %>%
  arrange(figure_number, coefficient, figure_panel) %>% 
  ungroup() %>%
  mutate(label = sub("\\([a-zA-Z0-9 /%]*\\)", "", label)) %>% 
  mutate(label = sub(" $", "", label)) %>% 
  mutate(label = sub("\\\\n", "", label)) %>% 
  mutate(figure_panel = toupper(figure_panel)) %>% 
  select(figure_panel, label, coefficient, p_less_0, p_above_0) %>%
  knitr::kable()
```

Next we will plot the model fits on top of the data. First we will need to create a data set with smaller increments on the predictors so that the plots are smooth:

```{r}
N <- 200
newdata <- group_by(d_long, response) %>%
  do({
    data.frame(ph_scaled = predict(poly(.$ph, 2), seq(min(.$ph), max(.$ph), length.out = N))[,1],
      ph = seq(min(.$ph), max(.$ph), length.out = N),
      ph_scaled2 = predict(poly(.$ph, 2), seq(min(.$ph), max(.$ph), length.out = N))[,2])
  })
newdata <- bind_rows(data.frame(newdata, nutrients = 0), data.frame(newdata, nutrients = 1))
```

Now make the predictions:

```{r}
ii <<- 0
p <- plyr::ldply(fits, function(x) {
  ii <<- ii + 1
  newdata_this_response <- as.data.frame(filter(newdata, response == names(fits)[[ii]]))
  pp <- posterior_linpred(x, newdata = newdata_this_response, re.form = NA)
  l <- apply(pp, 2, quantile, probs = 0.1)
  u <- apply(pp, 2, quantile, probs = 0.9)
  med <- apply(pp, 2, median)
  data.frame(l, med, u, newdata_this_response)
}) %>% as_data_frame()

pp <- inner_join(p, unique(select(d_long, response, log_transform, family)))
pp <- pp %>% group_by(response) %>%
  mutate(med = ifelse(log_transform | family == "count", exp(med), med)) %>%
  mutate(l = ifelse(log_transform | family == "count", exp(l), l)) %>%
  mutate(u = ifelse(log_transform | family == "count", exp(u), u)) %>%
  left_join(labels)
```

And finally make the plot:

```{r, fig.width=8.5}
d_long_plot <- d_long %>% group_by(response, nutrients, ph) %>%
  summarise(ph_scaled = ph_scaled[1], ph_scaled2 = ph_scaled2[1],
    value = ifelse(log_transform[1], exp(mean(log(value))), mean(value)),
    log_transform = log_transform[1],
    family = family[1], link = link[1]) %>%
  ungroup() %>%
  left_join(labels)

d_long_plot <- bind_rows(d_long_plot, d_ulva0) # for plotting only

cols <- c(brewer.pal(4, "Greys")[[3]], brewer.pal(4, "Blues")[[3]])
g <- ggplot(d_long_plot, aes(ph, value, colour = as.factor(nutrients))) + 
  geom_point(alpha = 0.3) +
  facet_wrap(~label, scales = "free_y") +
  geom_line(data = pp, aes(ph, y = med), lwd = 1) +
  geom_ribbon(data = pp, aes(ph, ymax = u, ymin = l, fill = as.factor(nutrients)), 
    alpha = 0.15, inherit.aes = FALSE) +
  theme_sleek() +
  labs(colour = "Nutrients", fill = "Nutrients", y = "Response value", x = "pH") +
  scale_color_manual(values = cols) + scale_fill_manual(values = cols)
g
# ggsave("figs/gg-fits.pdf", width = 10, height = 6)
```

# Separate figures for groups of responses 

```{r}
pp <- mutate(pp, nutrients_text = ifelse(nutrients == 0, "No Nutrients", "Nutrients added"))
d_long_plot <- mutate(d_long_plot, nutrients_text = ifelse(nutrients == 0, "No Nutrients", "Nutrients added"))

pp$label_wrap = stringr::str_wrap(pp$label, width = 99)
d_long_plot$label_wrap = stringr::str_wrap(d_long_plot$label, width = 99)

pp$label_wrap <- sub("\\\\n","\n", pp$label_wrap)
d_long_plot$label_wrap <- sub("\\\\n","\n", d_long_plot$label_wrap)

pp <- mutate(pp, 
  label_ordered = reorder(label_wrap, as.numeric(as.factor(figure_panel))))
d_long_plot <- mutate(d_long_plot, 
  label_ordered = reorder(label_wrap, as.numeric(as.factor(figure_panel))))

cols <- c(brewer.pal(4, "Greys")[[3]], brewer.pal(4, "Blues")[[3]])

plot_predictions <- function(figure_id, legend_position = c(.85, .96), 
  whole_numbers = FALSE) {
  
  dat <- filter(d_long_plot, substr(figure_panel, 0, 1) == figure_id)
  pp_data <-filter(pp, substr(figure_panel, 0, 1) == figure_id)
  these_responses <- unique(dat$response)
  
  max_ci <- group_by(pp, response) %>% 
    summarize(max_ci = max(u), min_ci = min(u))
  
  let <- dat %>% left_join(max_ci) %>% 
    group_by(response) %>%
    summarize(label_ordered = label_ordered[1], 
      x = max(ph) - 0.02, 
      range = diff(range(c(value, max_ci, min_ci))), 
      y = max(c(value, max_ci)) - range * 0.04) %>% 
    arrange(label_ordered) %>%
    mutate(panel = LETTERS[1:n()])
  
  g <- ggplot(dat, aes_string("ph", "value", colour = "nutrients_text")) + 
    geom_point(alpha = 0.45, aes(shape = nutrients_text)) +
    facet_wrap(~label_ordered, scales = "free_y") +
    geom_line(data = pp_data, aes_string("ph", y = "med"), lwd = 1) +
    geom_ribbon(data = pp_data, aes_string("ph", ymax = "u", ymin = "l", 
      fill = "nutrients_text"), 
      alpha = 0.15, inherit.aes = FALSE) +
    theme_sleek() +
    labs(colour = "Nutrients", fill = "Nutrients", y = "Response value", 
      x = expression(Acidification~(mean~seawater~pH[T]))) +
    scale_color_manual(values = cols) + scale_fill_manual(values = cols) +
    theme(legend.position = legend_position) +
    labs(color = "", fill = "", shape = "") +
    geom_text(data = let, aes(x = x, y = y, label = panel), inherit.aes = FALSE) +
    scale_shape_manual(values = c(19, 19)) +
    scale_x_reverse()
  
  if (whole_numbers) {
    g <- g + scale_y_continuous(breaks = seq(-20, 20, 2))
  }
  
  g
}
plot_predictions("3", legend_position = c(.75, .37))
ggsave("figs/fig-3.pdf", width = 4.6, height = 3.8)
plot_predictions("4", legend_position = c(.865, .7))
ggsave("figs/fig-4.pdf", width = 4.6, height = 3.8)
plot_predictions("5", legend_position = c(.87, .9))
ggsave("figs/fig-5.pdf", width = 6.9, height = 2.3)
plot_predictions("S", whole_numbers = TRUE, legend_position = c(.14, .85))
ggsave("figs/fig-S3.pdf", width = 6.9, height = 2.3)
```

Let's derive some values for the paper. 

```{r}
get_value <- function(response_value, ph_value) {
  xx <- filter(pp, response == response_value, ph >= ph_value - 0.001, nutrients == 1) %>% 
    arrange(ph)
  # print(xx$ph[[1]])
  xx <- xx[1, "med"]
  xx[[1]]
}

message("Epiphyte loading was X% longer at pH 7.45 than 7.86")
(x1 <- get_value("epiphyte_cm_shoot", 7.45))
(x2 <- get_value("epiphyte_cm_shoot", 7.86))
print(round(100 * (x1 / x2 - 1), 0))

message("Algal mass was approximately X-fold higher at 6.99 than 7.86")
(x1 <- get_value("algal_mass_g_dw", 6.99))
(x2 <- get_value("algal_mass_g_dw", 7.86))
round(x1/x2, 1)

message("There was x-fold higher diatom recruitment when nutrients were added:")
x <- as.data.frame(as.matrix(fits[["diatom_grazing"]]))[,"nutrients"]
print(round(exp(quantile(x, probs = c(0.025, 0.5, 0.975))), 1))

message("Elevated nutrient conditions were associated with X changes in mean seahare biomass ")
x <- as.data.frame(as.matrix(fits[["change_seahare_mass_g_fw"]]))[,"nutrients"]
print(round((quantile(x, probs = c(0.025, 0.5, 0.975))), 1))

message("Shoot mortality was x-fold higher when nutrients were added:")
x <- as.data.frame(as.matrix(fits[["dead_shoots"]]))[,"nutrients"]
print(round(exp(quantile(x, probs = c(0.025, 0.5, 0.975))), 1))

message("Shoot grazing was x-fold higher when nutrients were added:")
x <- as.data.frame(as.matrix(fits[["sh_grazing"]]))[,"nutrients"]
print(round(exp(quantile(x, probs = c(0.025, 0.5, 0.975))), 1))
```

```{r}
saveRDS(d_long, file = "data/generated/d_long.rds")
saveRDS(labels, file = "data/generated/labels.rds")
```
